+++ Sources/PortalSwift/PrivacyInfo.xcprivacy
